refactor: 상태 초기화 변수 MetadataManagementContent로 이동
